[6h30m] Add PCG basics and Dungeon Generation
Dungeon generation is not completely done, but the basics are working right now (it is possible to create rooms in different sizes and adjust the PCG accordingly). The new rooms can also be generated, but not loaded into the project yet and there are 1-2 bugs regarding the bounding boxes.
Already laid out the basics for the minimap as well, will be implemented soon.