[3h] Add code on HP-system, interaction system
Implementation of a basic health system (connected to Interfaces for
damage taking/healing up), as well as a basic interaction system and UI
for item pick up has been done.

For the Health I used a Blueprint Component to be able to potentially
reuse the code in AI Characters as well, without any changes
(hopefully).

On the interaction side, it was setup mostly for weapons, but because I
made it using an Interface it will be useable for stuff like doors or
buttons with little to no changes.

[There is probably a few more tweaks I did but forgot :)]