[8h] Add finishing touches and combine all systems functionalities
In this last commit I put together all the individual systems and fixed
bugs. So many BUGS. I will try to count here as many things as I can
remember that happened since the last commit:
- Add various VFX and SFX for the guns, the AI and explosions
- Fix shop to actually give out random guns from the (sadly only) two
available weapon types (it is very expandable though, since logic for
Projectile, Damage and the Gun itself is done with inheritance in mind)
- Add Heal Ups that spawn randomly when an enemy dies, will  heal one
heart (it is ANIMATED :D)
- Enemy Spawner is procedurally generated, so each room will have a
random quantity and positioning for the AI; all variables are exposed so
difficulty can be adjusted easily, or new AI types added;
- Add functionality for locking/unlocking rooms when enemies are still
in the room (just like BoI)
- Add OnDeath functionality for the player (level resets and all rooms
are regenerated)
- Fix Health Bar and Health values so they can work together (simplified
to 1 Heart = 1 HP)
- Fix Rocket Launcher > no self-damage, so rocket jumping is viable (and
fun!)
- Add dynamic material to enemies when getting hurt and messed around
with gun/arms materials for a different vibe
- ... most likely a few more things I forgot

Overall integrating all these systems was fun and challanging, a good
experience with trying to discover anothers' code and connect it with
yours, even though people can have completely different solutions to the
exact same and simple problem. ¯\_( ͡° ͜ʖ ͡°)_/¯